+++ tests/bazel.rs
@@ -1,89 +1,11 @@
-use keepsorted::{process_lines, process_lines_bazel};
+use keepsorted::process_lines_bazel;
 use std::io::{self};
 
 // Helper function to hide text-lines conversion.
-fn process_text(text: &str) -> io::Result<String> {
-    let lines: Vec<&str> = text.lines().collect();
-    let processed_lines = process_lines(lines)?;
-    Ok(processed_lines.join("\n"))
-}
-
-fn process_text_bazel(text: &str) -> io::Result<String> {
+fn process_bazel(text: &str) -> io::Result<String> {
     let lines: Vec<&str> = text.lines().collect();
     let processed_lines = process_lines_bazel(lines)?;
     Ok(processed_lines.join("\n"))
-}
-
-#[test]
-fn empty() {
-    let input = "";
-    let expected = "";
-    let result = process_text(input).unwrap();
-    assert!(result == expected, "Expected: {expected}\nActual: {result}");
-}
-
-#[test]
-fn single_letter() {
-    let input = "
-        a
-    ";
-    let expected = "
-        a
-    ";
-    let result = process_text(input).unwrap();
-    assert!(result == expected, "Expected: {expected}\nActual: {result}");
-}
-
-#[test]
-fn no_comment() {
-    let input = "
-        b
-        a
-    ";
-    let expected = "
-        b
-        a
-    ";
-    let result = process_text(input).unwrap();
-    assert!(result == expected, "Expected: {expected}\nActual: {result}");
-}
-
-#[test]
-fn simple_block() {
-    let input = "
-        # Keep sorted.
-        b
-        a
-    ";
-    let expected = "
-        # Keep sorted.
-        a
-        b
-    ";
-    let result = process_text(input).unwrap();
-    assert!(result == expected, "Expected: {expected}\nActual: {result}");
-}
-
-#[test]
-fn blocks_divided_by_newline() {
-    let input = "
-        # Keep sorted.
-        d
-        c
-
-        b
-        a
-    ";
-    let expected = "
-        # Keep sorted.
-        c
-        d
-
-        b
-        a
-    ";
-    let result = process_text(input).unwrap();
-    assert!(result == expected, "Expected: {expected}\nActual: {result}");
 }
 
 #[test]
@@ -102,7 +24,7 @@ fn bazel_block() {
             "b",
         ]
     "#;
-    let result = process_text_bazel(input).unwrap();
+    let result = process_bazel(input).unwrap();
     assert!(result == expected, "Expected: {expected}\nActual: {result}");
 }
 
@@ -130,7 +52,7 @@ fn bazel_block_with_comment() {
             # Trailing comment.
         ]
     "#;
-    let result = process_text_bazel(input).unwrap();
+    let result = process_bazel(input).unwrap();
     assert!(result == expected, "Expected: {expected}\nActual: {result}");
 }
 
@@ -158,7 +80,7 @@ fn bazel_blocks() {
             "x",
         ],
     "#;
-    let result = process_text_bazel(input).unwrap();
+    let result = process_bazel(input).unwrap();
     assert!(result == expected, "Expected: {expected}\nActual: {result}");
 }
 
@@ -200,7 +122,7 @@ fn bazel_blocks_select() {
             ],
         })
     "#;
-    let result = process_text_bazel(input).unwrap();
+    let result = process_bazel(input).unwrap();
     assert!(result == expected, "Expected: {expected}\nActual: {result}");
 }
 
@@ -228,6 +150,6 @@ fn bazel_order() {
             "@crate_index//:b",
         ]
     "#;
-    let result = process_text_bazel(input).unwrap();
+    let result = process_bazel(input).unwrap();
     assert!(result == expected, "Expected: {expected}\nActual: {result}");
 }
